+++ codewit/client/src/pages/ModuleForm.tsx
@@ -3,9 +3,9 @@ import Select, { MultiValue } from "react-select";
 import LanguageSelect from "../components/form/LanguageSelect";
 import TopicSelect from "../components/form/TagSelect";
 import ResourceSelect from "../components/form/ResourceSelect";
-import CreateButton from "../components/CreateButton";
-import ReusableTable from "../components/ReusableTable";
-import ReusableModal from "../components/ReusableModal";
+import CreateButton from "../components/form/CreateButton";
+import ReusableTable from "../components/form/ReusableTable";
+import ReusableModal from "../components/form/ReusableModal";
 import { toast } from "react-toastify";
 import { SelectedTag, Module } from "@codewit/interfaces";
 import { isFormValid } from "../utils/formValidationUtils";
